backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
pillow
numpy
pandas
gtts
easyocr
torch
torchvision
httpx
